% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/anovatxt.R
\name{anova_txt}
\alias{anova_txt}
\title{Add together two numbers}
\usage{
anova_txt(model_aov_4)
}
\arguments{
\item{x}{A number}

\item{y}{A number}
}
\value{
The sum of \code{x} and \code{y}
}
\description{
Add together two numbers
}
\examples{
add(1, 1)
add(10, 1)
}
