% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_2_interaction.R
\name{plot_2_interaction}
\alias{plot_2_interaction}
\title{Add together two numbers}
\usage{
plot_2_interaction(df, iv1, iv2, iv3, dv, continuous = c(F, T))
}
\arguments{
\item{x}{A number}

\item{y}{A number}
}
\value{
The sum of \code{x} and \code{y}
}
\description{
Add together two numbers
}
\examples{
add(1, 1)
add(10, 1)
}
