% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ttintertxt.R
\name{tt_inter_txt}
\alias{tt_inter_txt}
\title{Add together two numbers}
\usage{
tt_inter_txt(dv, iv, df, welch = c(F, T))
}
\arguments{
\item{x}{A number}

\item{y}{A number}
}
\value{
The sum of \code{x} and \code{y}
}
\description{
Add together two numbers
}
\examples{
tt_inter_txt(dv = "uptake", iv = "Type", df = CO2, welch = TRUE) # ...or welch = T
}
