% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ttmutxt.R
\name{tt_mu_txt}
\alias{tt_mu_txt}
\title{Add together two numbers}
\usage{
tt_mu_txt(dv, mu = 0, df)
}
\arguments{
\item{x}{A number}

\item{y}{A number}
}
\value{
The sum of \code{x} and \code{y}
}
\description{
Add together two numbers
}
\examples{
add(1, 1)
add(10, 1)
}
