% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/anova_txt.R
\name{anova_txt}
\alias{anova_txt}
\title{Add together two numbers}
\usage{
anova_txt(model_aov_4)
}
\arguments{
\item{afex::aov_4}{object}
}
\value{
anova output with inline for rmarkdown
}
\description{
Add together two numbers
}
\examples{
**example 1**
library(afex)
library(magrittr)
library(dplyr)
library(tibble)
library(tidyr)
data(mtcars)
data<-mtcars \%>\% mutate(ID = 1:nrow(.))
model<-afex::aov_4(mpg~ vs * am + (1|ID), data) # between-subject design
result<-anova_txt(model = model)
result$anova$txt

**example 2**
data(mtcars)
data<-mtcars \%>\% mutate(ID = rep(1:16, 2), intra = rep(c(0,1), each=16))
model<-afex::aov_4(mpg~ 1 + (intra|ID), data) # within-subject design
result<-anova_txt(model = model)
result$anova$txt
}
